refactor: refactor exceptions for message use in constants
- refactor exceptions for message use in constants
