+++ src/main/java/com/proyecto/proyecto/constant/Message.java
@@ -1,0 +1,17 @@
+package com.proyecto.proyecto.constant;
+
+public class Message {
+
+    public static final String USERNAME_NOT_FOUND = "The user with the following username: %d does not exist";
+
+    public static final String USERNAME_ALREADY_EXIST = "The following username: %d already exist";
+
+    public static final String USER_NOT_FOUND = "The user with id: %d not found";
+
+    public static final String BRAND_NOT_FOUND = "The brand with id: %d not found";
+
+    public static final String CATEGORY_NOT_FOUND = "The category with id: %d not found";
+
+    public static final String CLIENT_NOT_FOUND = "The client with id: %d not found";
+    
+}
