--- src/main/java/com/proyecto/proyecto/entity/AuthRequestDTO.java
@@ -1,9 +1,0 @@
-package com.proyecto.proyecto.model;
-
-import lombok.Data;
-
-@Data
-public class AuthRequestDTO {
-    private String username;
-    private String password;
-}
