--- src/main/java/com/proyecto/proyecto/entity/AuthResponseDTO.java
@@ -1,10 +1,0 @@
-package com.proyecto.proyecto.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class AuthResponseDTO {
-    private String token;
-}
