+++ src/main/java/com/proyecto/proyecto/enums/Role.java
@@ -1,0 +1,6 @@
+package com.proyecto.proyecto.enums;
+
+public enum Role {
+    USER,
+    ADMIN
+}
