+++ src/main/java/com/proyecto/proyecto/enums/State.java
@@ -1,4 +1,4 @@
-package com.proyecto.proyecto.model.enums;
+package com.proyecto.proyecto.enums;
 
 public enum State {
     ACTIVO,
