--- src/main/java/com/proyecto/proyecto/model/enums/Role.java
@@ -1,6 +1,0 @@
-package com.proyecto.proyecto.model.enums;
-
-public enum Role {
-    USER,
-    ADMIN
-}
