+++ src/main/java/com/proyecto/proyecto/service/user/UserServiceImpl.java
@@ -7,7 +7,8 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
-import com.proyecto.proyecto.model.enums.*;
+import com.proyecto.proyecto.constant.Message;
+import com.proyecto.proyecto.enums.*;
 import com.proyecto.proyecto.exception.NotFoundException;
 import com.proyecto.proyecto.exception.UsernameAlreadyExistsException;
 import com.proyecto.proyecto.model.User;
@@ -59,7 +60,7 @@ public class UserServiceImpl implements UserService {
     public Optional<User> checkIfUsernameExists(String username) {
         Optional<User> user = userRepository.findByUsername(username);
         if (user.isPresent()) {
-            throw new UsernameAlreadyExistsException("El nombre de usuario ya está en uso: " + username);
+            throw new UsernameAlreadyExistsException(String.format(Message.USERNAME_ALREADY_EXIST, username));
         }
         return user;
     }
@@ -78,7 +79,9 @@ public class UserServiceImpl implements UserService {
     @Override
     public User findByUsernameReturnToken(String username) {
         User user = userRepository.findByUsername(username)
-                .orElseThrow(() -> new UsernameNotFoundException("El usuario con ese username no existe: " + username));
+            .orElseThrow(
+                () -> new UsernameNotFoundException(String.format(Message.USERNAME_NOT_FOUND, username))
+            );
 
         String jwt = jwtProviderImpl.generateToken(user);
         user.setToken(jwt);
@@ -95,10 +98,10 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public Optional<User> findUserById(Long id) {
-        Optional<User> user = userRepository.findById(id);
+    public Optional<User> findUserById(Long id_user) {
+        Optional<User> user = userRepository.findById(id_user);
         if (user.isEmpty()) {
-            throw new NotFoundException("No se ha encontrado un el id: " + id);
+            throw new NotFoundException(String.format(Message.USER_NOT_FOUND, id_user));
         }
         return user;
     }
